Create musics, create albums with attached musics and genres
It is now possible to create musics, but not edit nor delete them (on purpose, for the sake of ease of work).
When creating albums, now every music is listed and it is now possible to check the music(s) that belong to that album. When on Details page of a specific album, every music attached to it is listed as well.
An album is associated ONLY with one artist/user (for the sake of ease of work, once again - although the relationship is still many-to-many between albums-artists).
Genres can now be listed ONLY - no edit, details and delete. Genres are ALWAYS the same, as it is seeded on the application.